+++ Assets/Scripts/Magic/Mana.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 
 public class Mana : MonoBehaviour
 {
-    public int Count;
+    public float Count;
 }
